Issue #226；#228

+++ CourseManagement/app/src/main/java/com/ftd/schaepher/coursemanagement/activities/ExcelDisplayActivity.java
@@ -221,9 +221,18 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
         AlertDialog.Builder mBuilder = new AlertDialog.Builder(ExcelDisplayActivity.this);
         LayoutInflater mInflater = ExcelDisplayActivity.this.getLayoutInflater();
         final View alertDialogView = mInflater.inflate(R.layout.dialog_excel_modify, null);
-        initRowWindowData(position, alertDialogView);
+
+        // 教师若已经提交则不能改；状态为2都不能改；
+        // 若是教师，且状态不为0，也不能改。
+        // 系负责人和教学办只有在状态为1时才能改
+        boolean doNotChange = hasCommitted || taskState.equals("2") ||
+                (identity.equals(ConstantStr.ID_TEACHER) && !taskState.equals("0")) ||
+                (!identity.equals(ConstantStr.ID_TEACHER) && taskState.equals("0"));
+
+        initRowWindowData(position, alertDialogView, doNotChange);
         mBuilder.setView(alertDialogView);
-        if (hasCommitted) {
+
+        if (doNotChange) {
             mBuilder.setNegativeButton("关闭", new DialogInterface.OnClickListener() {
                 @Override
                 public void onClick(DialogInterface dialog, int which) {
@@ -232,7 +241,7 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
             });
         } else {
             mBuilder
-                    .setPositiveButton("确认选择", new DialogInterface.OnClickListener() {
+                    .setPositiveButton("确认填写", new DialogInterface.OnClickListener() {
                         @Override
                         public void onClick(DialogInterface dialog, int which) {
                             EditText edtTxDialogTeacher =
@@ -241,27 +250,29 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
                                     (EditText) alertDialogView.findViewById(R.id.edtTx_dialog_from_to_end);
                             EditText edtTxDialogNote =
                                     (EditText) alertDialogView.findViewById(R.id.edtTx_dialog_note);
-
                             try {
-                                dbHelper.dropTable(commonTableName);
-                                dbHelper.changeTableName(tableName, commonTableName);
-
                                 TableCourseMultiline courseModify = new TableCourseMultiline();
                                 courseModify.setCourseName(excelListData.get(position).getCourseName());
                                 courseModify.setTimePeriod(edtTxDialogFromToEnd.getText().toString());
                                 courseModify.setRemark(edtTxDialogNote.getText().toString());
                                 courseModify.setTeacherName(edtTxDialogTeacher.getText().toString());
                                 courseModify.setWorkNumber(workNumber);
-                                dbHelper.update(courseModify);
-
-                                dbHelper.changeTableName(commonTableName, tableName);
+                                if (!identity.equals(ConstantStr.ID_TEACHER)) {
+                                    postOneLineToServer(courseModify);
+                                } else {
+                                    dbHelper.dropTable(commonTableName);
+                                    dbHelper.changeTableName(tableName, commonTableName);
+                                    dbHelper.update(courseModify);
+                                    dbHelper.changeTableName(commonTableName, tableName);
+                                    onResume();
+                                }
                             } catch (Exception e) {
                                 e.printStackTrace();
                             }
-                            onResume();
+
                         }
                     })
-                    .setNeutralButton("删除选课", new DialogInterface.OnClickListener() {
+                    .setNeutralButton("清空", new DialogInterface.OnClickListener() {
                         @Override
                         public void onClick(DialogInterface dialog, int which) {
                             try {
@@ -294,7 +305,7 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
     }
 
     // 设置弹窗的数据
-    private void initRowWindowData(int position, View v) {
+    private void initRowWindowData(int position, View v, boolean canBeChanged) {
         TextView tvDialogGrade = (TextView) v.findViewById(R.id.tv_dialog_grade);
         TextView tvDialogMajor = (TextView) v.findViewById(R.id.tv_dialog_major);
         TextView tvDialogNum = (TextView) v.findViewById(R.id.tv_dialog_sum);
@@ -321,7 +332,7 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
         edtTxDialogTeacher.setText(rowData.getTeacherName());
         edtTxDialogFromToEnd.setText(rowData.getTimePeriod());
         edtTxDialogNote.setText(rowData.getRemark());
-        if (hasCommitted) {
+        if (canBeChanged) {
             edtTxDialogTeacher.setFocusable(false);
             edtTxDialogTeacher.setEnabled(false);
             edtTxDialogFromToEnd.setFocusable(false);
@@ -338,12 +349,12 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
             case ConstantStr.ID_TEACHER:
                 menu.findItem(R.id.action_commit_task).setVisible(true);
                 break;
-            case ConstantStr.ID_DEPARTMENT_HEAD:
-                menu.findItem(R.id.action_commit_check).setVisible(true);
-                break;
-            case ConstantStr.ID_TEACHING_OFFICE:
-                menu.findItem(R.id.action_finish_task).setVisible(true);
-                break;
+//            case ConstantStr.ID_DEPARTMENT_HEAD:
+//                menu.findItem(R.id.action_commit_check).setVisible(true);
+//                break;
+//            case ConstantStr.ID_TEACHING_OFFICE:
+//                menu.findItem(R.id.action_finish_task).setVisible(true);
+//                break;
             default:
                 break;
         }
@@ -371,22 +382,22 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
                     }
                 }
                 return true;
-            case R.id.action_commit_check:
-                if (taskState.equals("2")) {
-                    showForbidCommitDialog("报课任务已结束，无法进行提交");
-                } else if (taskState.equals("1")) {
-                    showForbidCommitDialog("您已进行过提交，不能再次提交！");
-                } else {
-                    showCommitTaskDialog("是否提交审核", "一旦提交将不能再次修改报课信息！");
-                }
-                return true;
-            case R.id.action_finish_task:
-                if (taskState.equals("2")) {
-                    showForbidCommitDialog("报课任务已结束");
-                } else {
-                    showCommitTaskDialog("是否公开报课", "一旦公共将无法再进行修改！");
-                }
-                return true;
+//            case R.id.action_commit_check:
+//                if (taskState.equals("2")) {
+//                    showForbidCommitDialog("报课任务已结束，无法进行提交");
+//                } else if (taskState.equals("1")) {
+//                    showForbidCommitDialog("您已进行过提交，不能再次提交！");
+//                } else {
+//                    showCommitTaskDialog("是否提交审核", "一旦提交将不能再次修改报课信息！");
+//                }
+//                return true;
+//            case R.id.action_finish_task:
+//                if (taskState.equals("2")) {
+//                    showForbidCommitDialog("报课任务已结束");
+//                } else {
+//                    showCommitTaskDialog("是否公开报课", "一旦公共将无法再进行修改！");
+//                }
+//                return true;
             default:
                 return super.onOptionsItemSelected(item);
         }
@@ -441,6 +452,37 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
                 }).show();
     }
 
+    private void postOneLineToServer(final TableCourseMultiline course) throws IOException {
+        String cbTableName = "cb_" + tableName;
+        NetworkManager.postToServerAsync(cbTableName,
+                JsonTools.getJsonString(course), NetworkManager.INSERT_OR_UPDATE_CB_TABLE, new NetworkManager.ResponseCallback() {
+                    @Override
+                    public void onResponse(Response response) throws IOException {
+                        String result = response.body().string();
+                        Loger.w("oneLinePost", result);
+                        if (result.equals("true")) {
+                            dbHelper.dropTable(commonTableName);
+                            dbHelper.changeTableName(tableName, commonTableName);
+                            dbHelper.update(course);
+                            dbHelper.changeTableName(commonTableName, tableName);
+                            runOnUiThread(new Runnable() {
+                                @Override
+                                public void run() {
+                                    onResume();
+                                }
+                            });
+                        } else {
+                            sendToast("更新失败，请重试");
+                        }
+                    }
+
+                    @Override
+                    public void onFailure(Request request, IOException e) {
+
+                    }
+                });
+    }
+
     // 提交报课信息到服务器
     private void postToServer() {
         dbHelper.dropTable(commonTableName);
@@ -460,26 +502,27 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
                 e.printStackTrace();
                 sendToast("提交失败，请重新提交");
             }
-        } else {
-            List<TableCourseMultiline> commitData = dbHelper.findAll(TableCourseMultiline.class);
-            String cbTableName = "cb_" + tableName;
-            try {
-                String result = NetworkManager.postToServerSync(cbTableName,
-                        JsonTools.getJsonString(commitData), NetworkManager.INSERT_OR_UPDATE_CB_TABLE);
-                sendToast("提交成功");
-                if (identity.equals(ConstantStr.ID_DEPARTMENT_HEAD)) {
-                    TableTaskInfo task = new TableTaskInfo("1", tableName);
-                    dbHelper.update(task);
-                } else {
-                    TableTaskInfo task = new TableTaskInfo("2", tableName);
-                    dbHelper.update(task);
-                }
-                taskState = dbHelper.findById(tableName, TableTaskInfo.class).getTaskState();
-            } catch (IOException e) {
-                e.printStackTrace();
-                sendToast("提交失败，请重新提交");
-            }
         }
+//        else {
+//            List<TableCourseMultiline> commitData = dbHelper.findAll(TableCourseMultiline.class);
+//            String cbTableName = "cb_" + tableName;
+//            try {
+//                String result = NetworkManager.postToServerSync(cbTableName,
+//                        JsonTools.getJsonString(commitData), NetworkManager.INSERT_OR_UPDATE_CB_TABLE);
+//                sendToast("提交成功");
+//                if (identity.equals(ConstantStr.ID_DEPARTMENT_HEAD)) {
+//                    TableTaskInfo task = new TableTaskInfo("1", tableName);
+//                    dbHelper.update(task);
+//                } else {
+//                    TableTaskInfo task = new TableTaskInfo("2", tableName);
+//                    dbHelper.update(task);
+//                }
+//                taskState = dbHelper.findById(tableName, TableTaskInfo.class).getTaskState();
+//            } catch (IOException e) {
+//                e.printStackTrace();
+//                sendToast("提交失败，请重新提交");
+//            }
+//        }
         dbHelper.changeTableName(commonTableName, tableName);
     }
 
@@ -572,8 +615,8 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
 
     @Override
     public void onBackPressed() {
-        if (!hasCommitted) {
-            SimpleDialog commitTaskDialog = new SimpleDialog(ExcelDisplayActivity.this);
+        if (!hasCommitted && identity.equals(ConstantStr.ID_TEACHER)) {
+            final SimpleDialog commitTaskDialog = new SimpleDialog(ExcelDisplayActivity.this);
             commitTaskDialog.message("退出将清除所有修改")
                     .title("提示")
                     .positiveAction("确定")
@@ -582,7 +625,15 @@ public class ExcelDisplayActivity extends AppCompatActivity implements AdapterVi
                         public void onClick(View v) {
                             ExcelDisplayActivity.this.finish();
                         }
-                    }).show();
+                    })
+                    .negativeAction("取消")
+                    .negativeActionClickListener(new View.OnClickListener() {
+                        @Override
+                        public void onClick(View v) {
+                            commitTaskDialog.cancel();
+                        }
+                    })
+                    .show();
         }
     }
 }
